Failing test for 404 which becomes 406
See
http://jira.elifesciences.org:8080/browse/ELPP-1209?focusedCommentId=14753&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-14753

There is some commented out code that shows the code passing again from
filter() while trying to render an error page. Pay attention that if you
fix this once, the test will start passing again even if the code
changes because the rendering of the page is cached; I had to run `drush
cache:clear` to make the test red again.