feat: Update REST resources for collection items (#904)
* feat: Update REST resources for collection items

The REST resources for collection items have been updated to include changes in the 'AbstractRestResourceBase.php', 'CollectionItemRestResource.php' and 'PromotionalCollectionItemRestResource.php' files. The 'getArticleSnippet' method now accepts a second argument. The 'latestVersion' and 'minVersion' properties were also updated in the 'CollectionItemRestResource.php' file. Additional logic for handling 'reviewed-preprint' type content was also added.

* feat: update elife/api dependency version

The elife/api dependency version has been updated from "^2.21" to "*". The reference and URL in the composer.lock file have been updated to match the new version.

* feat: update elife/api version to ^2.24 in composer.json

This commit updates the version of the elife/api from '*' to '^2.24'. Also, the content-hash in the composer.lock file is updated accordingly.

* feat: update API versions in functional tests

Update the version of 'application/vnd.elife.collection+json' from 2 to 3 in UnrecognisedAcceptHeaderTest.php. Modify the versions in WrongVersionEndpointTest.php, set 'application/vnd.elife.collection+json' to 1 and 4, and change 'application/vnd.elife.promotional-collection+json' from 2 to 3.

* feat: update promotional-collection version in UnrecognisedAcceptHeaderTest

The version of 'promotional-collection' used in the UnrecognisedAcceptHeaderTest
has been updated from version 1 to version 2. This change ensures the test remains
up-to-date with the latest API version.