feat: Update Vagrant box and other enhancements (#943)
* feat: Update Vagrant box and other enhancements

- Switch from `geerlingguy/drupal-vm` to `geerlingguy/ubuntu2004` for Vagrant box
- Update Vagrant box version lock to `1.0.4`
- Change Vagrant IP to `192.168.56.9`
- Update Drupal major version to `9` in `config.yml`
- Update the composer install command in README.md
- Update vagrant box removal command in README.md

* feat: Update post provision scripts in config.yml

This commit removes the post_provision_scripts in the config.yml file.
The scripts were previously defined but are now temporarily disabled.

* feat: Temporarily disable goaws setup in README.md

The setup of goaws has been temporarily disabled as mentioned in the
updated README.md. A ticket has been created to restore this functionality.
Please note that this change should not impact most of the development work.