feat: Add Redis configuration for DDev setup (#944)
* feat: Add Redis configuration for DDev setup

This commit includes various changes for setting up Redis with DDev. The Redis configuration for DDev includes the addition of a redis manifest, config.yaml, docker-compose file, and multiple redis scripts, along with updates to the README.md for the setup instructions. Also, a redis-cli command file is added for running redis-cli inside the container. Further changes include adding a new Redis configuration in settings.ddev.redis.php and a setup script in setup-drupal-settings.sh. The changes ensure proper setup and configuration of redis and DDev together.

* feat: add localstack and aws-cli to docker-compose

This commit introduces localstack and aws-cli services in the docker-compose configuration. The localstack service is configured for sns and sqs services. The aws-cli service is set up to create various sns topics and an sqs queue.