feat: update league/html-to-markdown to 4.10.0
This commit updates the league/html-to-markdown library version to 4.10.0 in
composer.lock. The composer.lock file also indicates that PHPUnit has been
updated to support versions 4.8 and 5.7. In addition, there are changes to
the support and funding urls.

The "drupal/gin" package was also added to the list of packages in composer.lock.

Finally, in HtmlMarkdownSerializer.php, an adjustment has been made to decode
HTML special characters before processing preservation markup.